Create a common package to remove duplicated code
