+++ common/banner_ad_type.go
@@ -1,4 +1,4 @@
-package openrtb2
+package common
 
 // BannerAdType
 //
