+++ go.mod
@@ -1,9 +1,8 @@
-module github.com/onfocusio/openrtb-adagio
+module github.com/prebid/openrtb/v20
 
 go 1.16
 
 require (
 	github.com/onsi/ginkgo v1.16.1
 	github.com/onsi/gomega v1.11.0
-	github.com/prebid/openrtb/v20 v20.1.0
 )
